fix(download_stacked_image_and_mask): Set proper shape for stacked image.
Was using cols (bootm, top) to set the first array dimension which actually correspond to rows.

This was not causing issues when the bbox was a perfect square.
However, in some cases the lat lon to col row conversion could result in a rectangular shape.

This was causing the `ValueError: could not broadcast input array from shape (512,512,3) into shape (512,0,3)` error.